+++ packages/vitdoc-runtime/src/index.tsx
@@ -1,11 +1,11 @@
 import React, { Suspense, lazy } from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import { HashRouter, Route, Routes } from "react-router-dom";
 
 const Pure = lazy(() => import("./pure"));
 const Main = lazy(() => import("./main"));
 
-ReactDOM.render(
+createRoot(document.querySelector("#component-root")!).render(
   <HashRouter>
     <Routes>
       <Route
@@ -25,6 +25,5 @@ ReactDOM.render(
         }
       />
     </Routes>
-  </HashRouter>,
-  document.querySelector("#component-root")
+  </HashRouter>
 );