+++ packages/vitdoc-ui/src/utils/lang.ts
@@ -1,9 +1,5 @@
 export const isJsx = (lang: string) => /^[j|t]sx$/.test(lang);
 
-export function hasReact(str) {
-  return /ReactDOM\.render/.test(str);
-}
-
 export function isTypes(str) {
   return /renderType\$\(.+\)/.test(str);
 }