+++ packages/vitdoc-ui/src/utils/react-render.ts
@@ -1,0 +1,6 @@
+import { ReactNode } from "react";
+import { createRoot } from "react-dom/client";
+
+export function ReactRender(ele: ReactNode, root: HTMLElement) {
+  createRoot(root).render(ele);
+}