+++ packages/vitdoc/src/utils/lang.ts
@@ -1,8 +1,6 @@
 export const isJsx = (lang: string) => /^[j|t]sx$/.test(lang);
 
-export function hasReact(str) {
-  return /ReactDOM\.render/.test(str);
-}
+
 
 export function isTypes(str) {
   return /renderType\$\(.+\)/.test(str);