+++ packages/vitdoc-ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vitdoc/ui
 
+## 1.3.5-beta.0
+
+### Patch Changes
+
+- react 18
+
 ## 1.3.4
 
 ### Patch Changes